Implemented flash sessions
This is a fairly simple implementation of flash sessions. It reuses as
much of the infrastructure put in place by signed sessions as possible.
We're currently only resetting flash session on non failure status codes
which isn't correct and is fairly hacky so i'll have to create another
issue for ironing that out later. This just gets it in a workable state
for now.

A decent portion of this PR was just moving the http responses in a more
consolidated place so they'd be easier to manage from now on.