Very slight tweaks to front object
Was playing with the front object last night. Came to the conclusion
that things that are user overidable should be methods not values. So
i moved all the vals to defs. Also the main method should always be the
same because the user should drive their configuration from the methods
not from main, so i moved the main method to the front object trait
so it's just always defined.