Implemented Specs2
I implemented Specs2 into the framework and implemented a very small
set of some tests. There will be plenty of tests to write but before
we go down that road i'd like to get jenkins/travis set up to run
builds on commit and i'd also like to setup code coverage reports so
we can track our status on a per commit basis on how well we're advancing
in the testing effort.